feat: add post-body jiraLocation
This comes with a bit of complexity. Previously we were automatically appending a space to the jira
part. However by moving it here, we actually wish to have the space at the start. This then adds
complications with the jiraPrepend and jiraAppend options, and thus causes a breaking change.

BREAKING CHANGE: Existing jiraAppend options will need to add an extra space to maintain
functionality